Script for filling county data
This script will take the input file (expected to have a column for Origin County already) and fill based on Origin City and Origin State. Keeps temporary track of all already found counties as to not have to utilize the geopy library every time. This saves a lot of time when working with large sets of data where many cities are the same.